+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.5.6
+
+- Catch and log exceptions thrown in methods `rollbackStore()` and `rollbackUpdate()`.
+
+## 3.5.5
+
+- Updated methods `rollbackStore()` and `rollbackUpdate()`.
+
 ## 3.4.1
 
 - Fixed typo